+++ src/globals.rs
@@ -1,5 +1,6 @@
 use mlua::prelude::*;
-use crate::*;
+use crate::prelude::*;
+use crate::{require, std_io_output};
 
 pub fn error(_luau: &Lua, error_value: LuaValue) -> LuaValueResult {
     wrap_err!("message: {:?}", error_value.to_string()?)
