+++ src/std_fs/dirlib.rs
@@ -2,9 +2,7 @@ use std::fs;
 use std::io;
 use std::path::PathBuf;
 use mlua::prelude::*;
-use crate::LuaMultiResult;
-use crate::{wrap_err, colors, LuaValueResult, TableBuilder};
-use crate::ok_table;
+use crate::prelude::*;
 use super::entry::wrap_io_read_errors;
 use super::validate_path_without_checking_fs;
 use super::{directory_entry, entry, validate_path};
