+++ src/std_fs/entry.rs
@@ -4,8 +4,8 @@ use std::io::ErrorKind;
 use std::path::Path;
 use std::path::PathBuf;
 use mlua::prelude::*;
-use crate::{std_time, ok_table};
-use crate::{wrap_err, LuaValueResult, LuaEmptyResult, table_helpers::TableBuilder, colors};
+use crate::prelude::*;
+use crate::std_time;
 use copy_dir::copy_dir;
 
 #[cfg(unix)]
