+++ src/std_fs/file_entry.rs
@@ -1,6 +1,5 @@
 use mlua::prelude::*;
-use crate::{LuaValueResult, LuaEmptyResult, wrap_err};
-use crate::{colors, table_helpers::TableBuilder};
+use crate::prelude::*;
 use crate::std_fs::entry::{self, wrap_io_read_errors, wrap_io_read_errors_empty, get_path_from_entry};
 use std::cell::RefCell;
 use std::fs::{self, OpenOptions};
